backend/pyproject.toml
[tool]

[tool.poetry]
name = "backend"
description = "Backend"
version = "0.1.0"
authors = ["DTTA AMS Developers <dtta-ams-dev@catalyst.net.nz>"]

[[tool.poetry.source]]
name = "package proxy"
url = "https://cat-prod-buildproxy.dmz.wgtn.cat-it.co.nz/root/pypi/+simple/"
priority = "default"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-configurations = "*"
psycopg2-binary = "*"
gunicorn = "*"
sentry-sdk = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-django = "*"
pytest-subtests = "*"
mypy = "*"
black = "*"
flake8-bugbear = "*"
pylint = "*"
ipdb = "*"

[tool.black]
line-length = 120
